T442: minor cleanup
